customer-review-stream/services/api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8
prometheus-client==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1
